chore: extract session logic to Sessions class, restore nested plugins
