+++ Sources/Amplitude/Plugins/AmplitudeDestinationPlugin.swift
@@ -55,6 +55,8 @@ public class AmplitudeDestinationPlugin: DestinationPlugin {
             pipeline: pipeline!
         )
         pipeline?.start()
+
+        add(plugin: IdentityEventSender())
     }
 
     public override func execute(event: BaseEvent?) -> BaseEvent? {
