+++ Sources/Amplitude/Plugins/IdentityEventSender.swift
@@ -1,0 +1,20 @@
+//
+//  IdentityEventSender.swift
+//
+//
+//  Created by Hao Yu on 11/15/22.
+//
+
+internal class IdentityEventSender: Plugin {
+    var type: PluginType = .before
+
+    var amplitude: Amplitude?
+
+    func setup(amplitude: Amplitude) {
+
+    }
+
+    func execute(event: BaseEvent?) -> BaseEvent? {
+        return event
+    }
+}
